Check if shim-x64 package is present in the system
In PR#553 we removed a condition from the shim-x64 special case workaround
where it was only supposed to handle UEFI systems, and now we are handling BIOS
as well. The problem is that in the BIOS systems, this package may not be
installed at all.

This commit introduces a quick check to validate if the package is present on
the system before trying to backup it's protection file under
`/etc/yum/protected.d`.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>