[RHELC-635] Call integration tests through GitHub Actions (#574)
* Call integration tests through GitHub Actions

We are introducing a new GitHub Action to run the integration tests with the Testing Farm. This is a first step towards moving away from running the tests through Packit which is unable to access GitHub Secrets (while GitHub Action is). Our current implementation of accessing secrets through a file behind Red Hat VPN on an internal GitLab instance is not secure enough and the GitLab server is refusing to serve the number of requests we do.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>