Improve the internet connection check
This check was introduced in PR#458 as a quick way to validate if the internet
connectionw as up or not. The way that it was introduced was not very robust,
as it was checking for a public Google DNS that may be blocked on some
machines/servers as not essential.

This commit tries to improve it by, instead of checking Google Public DNS, we
check a static file hosted by Red Hat. That should be a little more robust
since `*.redhat.com` is supposed to be reachable from the customer servers when
running convert2rhel.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>