chore: build GHCR images on main
Forks don't have permission to update the image. We also don't need to update the image for every commit in a PR, intead we can simply update the images on push to main and use that image

Caveat: with this update we won't be able to update the images and have the workflows reflect that in the PRs themselves. We can utilize workflow_dispatch together with reusable workflows and environment variables to manually trigger this in the future if we want to