Fix permissions check for certificates.
We check whether we can access the certificate files shipped with
convert2rhel prior to installing them.  The check for the directory they
live in was checking whether the directory was readable and writable.
We don't write to that directory, though, so the check shuld be whether
the directory is readable and executable.

Note that convert2rhel should always be run by root so this bug won't be
encountered in normal use.