Dump the breadcrumbs into the RHSM facts format
Added a new function to the breadcrumbs class to dump the current view of the
breadcrmubs data into the RHSM custom facts format (single level json).

First, we refactored the way we sanitaze the CLI parameters in the breadcrumbs
by just using the `hide_secrets()` function (that is now in the `utils.py`
module) to be our main source of sanitization instead of the old function that
was way more complex and did the same thing as the `hide_secrets()`.

We also introduced a few changes to the breadcrumbs class to make it simpler
and have only one method to call to determine if the data collection is
finished or not.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>