Delete greetings workflow
It is mostly bloat, and we don't have outside contributors. If they do most of the information in the comment is unrelated to them anyway. As the maintainers of the tool we already know what is going on and have it documented elsewhere

If we need to add information we can do so in PR templates