Change the way we create the transaction_handler object
We were seeing a problem of recursion when we tried to import the class
in a python shell (and on some unit tests on centos8), thus, leaving a
problem with the code that it could happen anytime, anywhere.

This commit introduces changes to make it not recursive anymore, and
create the transaction_handler in a different way.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>