Remove epoch from yum NEVRA before removing pkg
The rpm command does not parse well the yum notation of an epoch in a pkg
NEVRA string when removing a pkg.
The yum notation: "epoch:name-version-release.arch"

Also, for debug purposes, print a yum output when the transaction
dependency resolution fails.

Also, fix the test_single_yum_transaction() integration test that expected
output strings in a wrong order.

Also, fix the test_resolve_yum_problematic_dependencies() unit test that
wasn't updated with the change of backup.remove_pkgs() in one of the
previous commits.