Test: Remove oldest kernel to prevent conflict
The test_system_not_updated fails on CentOS Linux 8.5 with an dnf transaction
calculation error if the maximum number of kernels that can be installed has
been reached:
"""
package kernel-4.18.0-348.23.1.el8_5.x86_64 requires
kernel-core-uname-r = 4.18.0-348.23.1.el8_5.x86_64, but none of the providers
can be installed
"""
To us now the only known solution is to remove the oldest kernel before proceeding.

Also, increase the timeout for waiting for the transaction validation.
The tests were failing for reaching the original timeout.