Fix failing schema validation
add anyOF directive and cover null values

source_os is set to "null" when we fail before early_collectio() has
been run.  We should talk about whether to run "early_collection" in
rollback if it hasn't been done already but that's a separate issue to
be determined some other time.

Signed-off-by: Daniel Diblik <ddiblik@redhat.com>