Fix sub-man installation when json-c.i686 installed
When json-c.i686 was installed it caused the conversion to fail at the
step of installing the pre-downloaded subscription-manager packages.

In case the json-c.i686 is installed we need to download it together with
its x86_64 companion. The reason is that it's not possible to install
a 64-bit library that has a different version from the 32-bit one.

Also, fix the test_single_yum_transaction integration test that
incorrectly decided whether the system uses yum or dnf.

Also, fix the test_system_not_updated integration test where we need to
remove not only the oldest kernel package on CentOS Linux 8 but all the
kernel* packages of the same version as the kernel itself.