Allow but warn about deprecated env var
We've renamed CONVERT2RHEL_UNSUPPORTED_VERSION to
CONVERT2RHEL_ALLOW_OLDER_VERSION.  Allow the old env var name to be used
but emit a deprecation warning in the log.

Add unittests to show the deprecation message is emitted when we use the
old env var rather than the new one to allow the conversion to continue.