[RHELC-637, RHELC-742] Improve internet connection check (#587)
* Improve the internet connection check

This check was introduced in PR#458 as a quick way to validate if the internet
connectionw as up or not. The way that it was introduced was not very robust,
as it was checking for a public Google DNS that may be blocked on some
machines/servers as not essential.

This commit tries to improve it by, instead of checking Google Public DNS, we
check a static file hosted by Red Hat. That should be a little more robust
since `*.redhat.com` is supposed to be reachable from the customer servers when
running convert2rhel.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

* Add integration tests

Adding integration tests to this feature, since in the PR (#458) when it
was introduced, there were no integration tests.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

* Fix typo

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

* Remove unused import

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

* Apply suggestions from code review

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>
Co-authored-by: kokesak <43987098+kokesak@users.noreply.github.com>
Co-authored-by: Toshio Kuratomi <a.badger@gmail.com>

* Apply fixes for the tests

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

* Update convert2rhel/systeminfo.py

Co-authored-by: Eric Gustavsson <eric@spytec.se>

* Improve messages for network connection check

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

* Apply suggestions from code review

Co-authored-by: Eric Gustavsson <eric@spytec.se>

* Improve messages for network connection check

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

* Tweak integration tests

Modify expected output for both tests.
Align metadata and testfiles to the proposed tier0 modifications.

Signed-off-by: Daniel Diblik <ddiblik@redhat.com>

* Modify integration test to overwrite a line in the resolv.conf

We were appending the `nameserver 127.0.0.1` to the /etc/resolv.conf
file instead of overwriting it, as it was not considering the system to
be "offline"

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>
Signed-off-by: Daniel Diblik <ddiblik@redhat.com>
Co-authored-by: kokesak <43987098+kokesak@users.noreply.github.com>
Co-authored-by: Toshio Kuratomi <a.badger@gmail.com>
Co-authored-by: Eric Gustavsson <eric@spytec.se>
Co-authored-by: Daniel Diblik <ddiblik@redhat.com>