Fix the format of message about unavailable kernel modules.
In #469, one of the cleanups ended up calling `"\n".join()` on a list of
kernel modules twice.  The first time changes the list into a string
with each entry separated by a newline but the second time changes the
string into another string with each chatacter separated by a newline.
Fixed that by removing the second `str.join()`.

This also refactors away a call to `map()` and use of `lambda`.  In
modern python, any uses of `map()` or `filter()` can be replaced with
either a generator expression or list comprehension.

Related to #469