Move system info logs to happen after the data collection
After we introduced the collection of breadcrumbs (#612) as a RHSM facts, we
started to use some internal (now public) functions from the
systeminfo.py module, which at first, those functions were responsible
for logging out some system information, such as: Name, Version, Config
file used, and so on... That change caused some duplicated lines in the
Convert2RHEL log output.

This commit fixes this by moving the logs to the `resolve_system_info`
function instead of placing the logs in each of the facts collection.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>