Update repoquery command for latest kernel check
What's happening is when testing on the command line, those quotes are
interpreted by the shell before they get passed to repoquery.  Inside of
python, since we're using subprocess with shell=False (for security), we need
to get rid of the quotes ourselves.

Updated the unit_tests related to the `is_loaded_kernel_latest` to reflect the
command that was changed, this way, we mock the `run_subprocess` properly.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>