Run the sub_man_rollback integration test on CentOS-8.4
This test exposes a failure trying to reinstall subscription-manager
packages but only when we haven't downloaded the yum metadata for the
convert2rhel repositories previously.  In current HEAD, the metadata is
downloaded when we backup the packages prior to reinstalling, *unless*
we're doing a CentOS-8.4 EUS release.  On the EUS release, we end up
using hte hardcoded repos that are shipped with convert2rhel in the
backup step:
https://github.com/oamg/convert2rhel/blob/50216c1ee8ff6c4f1c7d3f165a206d0dee93bb81/convert2rhel/backup.py#L345

Enabling this test on CentOS-8.4 will ensure that the error has been
fixed with this workaround.