[RHELC-731] feat: log command used (#638)
* refactor: move hide_secrets to utils

As we want to utilize hide_secrets when logging the convert2rhel command used, refactoring it to be within utils file is more appropriate

* feat: log command used

To better debug issues, we log the command used so it shows in the logfiles. To make sure no secrets are shown we hide them

* fix: bad usage of format

log_command_used test now properly imports

* fix: correct order of hiding secrets

* test: join expected_command in test

* Add verification test

* verify the passed command is logged at the top of the log file
* verify the secret values are obfuscated
* sanity check the logfile is created

Signed-off-by: Daniel Diblik <ddiblik@redhat.com>

* fix: properly hide secrets from params

Within the convert2rhel parameters there was shorthands for
password and activation key. We did not properly escape
these within `hide_secrets()`

* docs: fix typo in docstring

Co-authored-by: Rodolfo Olivieri <rolivier@redhat.com>

* test: update mocked credentials

It used terms which are used within controverial circles

* Assert just second line

The actual command gets printed to a second line of the logfile. Asert
just the second line, to avoid invalid assertion errors.

Signed-off-by: Daniel Diblik <ddiblik@redhat.com>

* Fix assertion

Signed-off-by: Daniel Diblik <ddiblik@redhat.com>

* test(hide_secrets): test the default secret_args

* fix(hide_secrets): fix pre-commit mistake

* Fix rebase and unit_tests for hide_secrets

After rebasing from main, there was a couple of conflicts regarding the
files present here.

I went ahead and did the rebase and fixed a couple of unit_tests to use
the ones that already exists from main.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

Signed-off-by: Daniel Diblik <ddiblik@redhat.com>
Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>
Co-authored-by: Daniel Diblik <ddiblik@redhat.com>
Co-authored-by: Rodolfo Olivieri <rolivier@redhat.com>