[RHELC-744] Install the redhat-uep.pem certificate if it is needed. (#642)
* Install the redhat-uep.pem certificate if it is needed.

* We need this certificate for the convert2rhel repo if the user installed
  using the convert2rhel.repo file:

  https://ftp.redhat.com/redhat/convert2rhel/8/convert2rhel.repo

  This is a workaround for https://issues.redhat.com/browse/RHELC-744
  When we make uninstalling and reinstalling the RHEL
  subscription-manager packages a single step process, we will no longer
  need this as there will always be a redhat-uep.pem available from an
  rpm package.

* Fix missing repo certificate for RHEL7.  On RHEL7, the missing
  redhat-uep.pem certificate causes additional issues.  When we check
  for the latest kernel, we do not get an error but a warning is printed
  in the repoquery output which makes our screenscraping of the output
  fail.

* Unittests for ca-file workaround

* Run the sub_man_rollback integration test on CentOS-8.4

This test exposes a failure trying to reinstall subscription-manager
packages but only when we haven't downloaded the yum metadata for the
convert2rhel repositories previously.  In current HEAD, the metadata is
downloaded when we backup the packages prior to reinstalling, *unless*
we're doing a CentOS-8.4 EUS release.  On the EUS release, we end up
using hte hardcoded repos that are shipped with convert2rhel in the
backup step:
https://github.com/oamg/convert2rhel/blob/50216c1ee8ff6c4f1c7d3f165a206d0dee93bb81/convert2rhel/backup.py#L345

Enabling this test on CentOS-8.4 will ensure that the error has been
fixed with this workaround.