Remove shim-x64 special workaround
Since the PR#528 landed in upstream, we included the shim-x64 package as
part of the excluded packages for Oracle Linux 7, meaning that we remove
this package before doing any operation in Convert2RHEL, thus, we don't
require the special workaround any more as this package will not be
present on the system when it reaches that part of the code.

The special workaround we had was to remove the protected file that the
shim-x64 package places on /etc/yum/protected.d when the package is
installed, since we remove it, the protection file is not present any
more.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>