Add missing integration tests for single yum transaction
While testing convert2rhel manually in a vagrant box, I realized that the
conversion stopped in the yum transaction validation phase, generating a
traceback and not doing a proper rollback after the error appeared.

Since the transaction validation is supposed to trigger the rollback in case of
any failures, it was strange that the tool was not behaving the way it should.

As part of #674 that introduced the fix, we left out the integration tests that
were supposed to cover this scenario.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>