Improve repoquery parsing for loaded kernel check
The repoquery parsing for the loaded kernel check was only filtering out
a single line that was present in the output, we were filtering this out
and removing it with the intention to clean the output and let only the
relevant output of repoquery to be parsed, however, we discovered that
different text can appear in the output, thus, leaving us with a broken
workflow again.

The solution to fix this problem was simply adding an identifier to the
repoquery output, this way we can filter out only the lines we want and
ignore the rest (logging them as debug messages in case we miss anything
important from that line).

In this commit I'm introducing a small refactoration for that function
and it's unit_tests as well.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>