chore: add devcontainer (#672)
Since convert2rhel is a volatile tool to run on our own machines without
prior understanding that the tests we are running are mocked out
correctly. We may break our machine. It is better to run within a
container but we then get into issues where debugging, testing, and so
on isn't as integrated with existing tools as we'd like.

This is the first step to adding devcontainer using VSCode Remote
Development which will show VSCode as normal but within a container,
making it much easier to run tests and debug