Remove black args from pre-commit-config file
We had pre-commit-config settings the args for black and the
pyproject.toml file as also setting some properties for the tool.

It seems that black override those properties from pyproject.toml since
we used the CLI args for it, but in any case, we should rely only on one
configuration file to tell the tool what property it should override.

Also removed the unnecessary property in pre-commit-config for `isort`.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>