Add progress indicators for DNF/YUM
This commit introduces the callbacks classes for the progress indicators
that we will be showing for the user during a few different phases in
our code.

The callbacks will called during these phases:

* For DNF: Dependency Solver, Package Download and Transaction Progress
* For YUM: Package Download and Transaction Progress

We don't have a dependency solver callback for yum, because the yum code
base does not provide any base class for us to override and iterate
over, thus, we would require to copy and paste their own dependency
resolver callback and deal adjust the code with the things we want and
the ones we don't care. Since those classes are highly tied to the
`base` usage, it is possible that we could mess with something and lead
to obscure errors, thus, we are not using any dependency solver callback
for YUM.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>