[KAIZEN] Make pylint logs usable (#702)
* Make pylint logs usable.

The goal is to make pylint output usable for finding regressions in the
agreed upon style.

* Change the pylint config to disable every check that currently fails.
  We will gradually go over these, adding back any that we care about and
  are willing to fix.

* Change the Makefile and github workflow to call pylint directly
  instead of the wrapper script which on fails on pylint Error level or
  higher.

* Add pylint as a dependency in pre-commit

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

* Change order of pre-commit hooks

Changed the order of pre-commit hooks to make the priority output appear
first in the logs (both locally and on pre-commit.ci), this way, we can
see if anything goes wrong during CI execution.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

* style: fix pylint errors for import and var usage

The pylint errors showed import error for os.path, changing this to just
from os import path fixed that issue. As for the variable usage, we
accessed setwinsize before it was initialized, but since it's set in
pexpect.spawn we just tell pylint to ignore it

* Fix pylint remaining errors

There were a few errors related to import relatives in a few source
files.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

* Remove the run_lint script and the lint-errors Makefile target.

We're now using the pylint config file to silence all the remaining
messages so we do not need the script which only fails if there were
errors.

* Upgrade required pylint to remove a false positive on `import os.path`.

pylint-2.13 was marking `import os.path` as an error:
  [E0611 no-name-in-module] No name 'path' in module 'os'

Upgrading to pylint-2.15 (.0 or higher) fixes the issue.

Also changing `import os` back to `import os.path` where appropriate.

* Reduce output from tmt.

Setting verbose: False for the tmt hook means that it will only output
if the test fails.  This should make it so we don't get tmt output that
causes pre-commit.ci to truncate useful information in other hooks if
only the other hooks are failing.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>
Co-authored-by: Rodolfo Olivieri <rolivier@redhat.com>
Co-authored-by: Eric Gustavsson <egustavs@redhat.com>