Add a later check to warn the user about kernel boot files
We added a few later checks to verify and warn the user if there is any
errors with both `initramfs` `vmlinuz` files on their system after the
conversion.

This check has the intention to at least inform the user what are the
steps that they need to run in order to fix that problem.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>