Change the label for FILE output to DEBUG.
The contents of FILE are the same as DEBUG.  The only difference is that
FILE outputs to the log file whereas DEBUG goes to the terminal and the
log file.  This means, in the log file, when we run convert2rhel with
the --debug flag, the log messages are labelled with DEBUG.  When we run
without --debug, the same messages end up in the log file but they have
the label FILE.  Changing the label of LogLevelFile to DEBUG will keep
the label the same for both of these cases.