Move adding of NullHandler to the root logger.
r0x0d added initialize.py for steps that need to be taken early to
initialize various subsystems.  Adding of the NullHandler to the root
logger (to avoid some libraries initializing themselves in a way that
ends up outputting messages multiple times) needs to be done very early.
So moving it to initialize.py is the logical choice.

Also removed a piece of code related to python 2.6 since we're dropping
support for that for the convert2rhel-1.2 release.