Rename the function to disable_root_logger()
Rodolfo correctly pointed out that calling this function
initialize_logging() begged the question, why aren't we also calling
main.initialize_logging()?  What does this do that the other one
doesn't?

Renaming this makes it clear that the reason this function exists is to
disable the root logger rather than to do all of the logging
initialization tasks.

Note: We don't want to do all of the logging initialization here because
the purpose of this file is to initialize anything that must happen
very early, before trivial things like importing certain libraries
happens.  If we performed the whole logging initialization (instead of
just this one targetted task), there is a danger that future changes
to the other portions of logging initialization would pull in additional
libraries before we were ready for them.  In general, we want to put the
bare minimum of code and as few dependencies as possible in this early
initialization step.