Add unit tests for later checks
Added some unit tests to assert that the correct path is being executed while
we are checking for the kernel boot files. The tests added here tries to assert
that both individual functions (checking for initramfs and vmlinuz) are called
correctly and going through the correct paths, as well, checking that the
workflow for the main check function is also doing what is supposed to.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>