Rename unavailable kmod inhibitor envvar
Before we do an upstream release we can still change the
CONVERT2RHEL_ALLOW_UNCHECKED_KMODS environment variable used for
overriding the check that inhibits the conversion when kernel modules
not available in RHEL are loaded.

It is not clear what we mean by the word UNCHECKED. The word UNAVAILABLE
is more understandable.

Also, the use of "first-party" kernel modules is makes it unclear what
we mean by that.