Fix "missing boot files" test (#748)
On CentOS and Oracle Linux 7 the test was failing because we were
renaming the vmlinuz and initramfs files in order to back them up, but
we left them in the /boot/ folder. That caused the grub2-mkconfig to
pick them up and generated a new grub configuration using these renamed
boot files.

This change does not back up the boot files but instead simply removes
them, simulating better the core test scenario of this integration test
which is the kernel rpm scriptlet failure to copy/generate these boot
files. After verifying that convert2rhel detected this situation of
missing boot files correctly, the test goes on to do what the user
instructed to do in such a case - that is reinstalling the kernel and
re-runing grub2-mkconfig.

Fixes https://github.com/oamg/convert2rhel/pull/721.