[RHELC-724] Move adding of NullHandler to the root logger. (#732)
* Move adding of NullHandler to the root logger.

r0x0d added initialize.py for steps that need to be taken early to
initialize various subsystems.  Adding of the NullHandler to the root
logger (to avoid some libraries initializing themselves in a way that
ends up outputting messages multiple times) needs to be done very early.
So moving it to initialize.py is the logical choice.

Note: We don't want to do all of the logging initialization here because
the purpose of this file is to initialize anything that must happen
very early, before trivial things like importing certain libraries
happens.  If we performed the whole logging initialization (instead of
just this one targetted task), there is a danger that future changes
to the other portions of logging initialization would pull in additional
libraries before we were ready for them.  In general, we want to put the
bare minimum of code and as few dependencies as possible in this early
initialization step.

Also removed a piece of code related to python 2.6 since we're dropping
support for that for the convert2rhel-1.2 release.