Disallow python3X-rpm-macros when building rpm
The modular pkgs python38-rpm-macros or python39-rpm-macros (which provide the
python-modular-rpm-macros capability) can be installed together with the
non-modular python3-rpm-macros package. And macros from the modular packages
override macros from the non-modular package.

Having a modular macro pkg installed would expand the %{python3_pkgversion}
we use in the .spec file to 38 or 39 instead of the expected 3. We use
%{python3_pkgversion} to get dependency package names, like
python%{python3_pkgversion}-six in order to get python3-six. Even though
python38-six and python39-six are available in RHEL 8 AppStream, the default
RHEL 8 python version is 3.6 and we want to build the rpm for that.

Before this change, with python38-rpm-macros installed:
```
error: Failed build dependencies:
    python38-six is needed by convert2rhel-1.2-1.el8.noarch
```

After this change, with python38-rpm-macros installed:
```
error: Failed build dependencies:
    python38-six is needed by convert2rhel-1.2-1.el8.noarch
    python-modular-rpm-macros conflicts with convert2rhel-1.2-1.el8.noarch
```