Handle SIGINT during conversion
Fix a bug where we couldn't issue a Ctrl + C (SIGINT) to Convert2RHEL
and start a rollback immediately.

This commit introduces the ability to run some functions as a child
process to the main process in order to be able to make `rpm` and `yum`
install their signal handling logic inside the child process rather than
the main process and break our workflow.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

Apply suggestions from code review

Co-authored-by: Preston Watson <104600742+pr-watson@users.noreply.github.com>