Fix running pylint in a container
Running `make lint` was failing due to not having pylint available in
the CentOS Linux 8 container (we run the linting in just this one
container).

Pylint writes logs to the home folder of the user running the pylint -
so this change also allows creating the user home folder by removing
`-r` from the useradd command.