Add environment variable for experimental conversion analysis
Introducing an environment variable that will enable the experimental
conversion analysis with report and rollback.

The intention of this environment variable is to act as a place holder
before we introduce the refactoration of our CLI to deal with
subcommands, in which, that environment variable will be converted to a
subcommand.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>