[RHELC-940] Convert Pre Ponr checks to the action framework (#764)
* Convert Pre Ponr checks to the action framework

Converting the pre_ponr checks to be part of the action framework.

This commit introduces an complete rewrite of the
`pre_ponr_conversion()`. All functions placed inside that method were
converted to an action class to be executed and handled inside the
action framework.

There is quite a few things left to be done, but because of time
constraint, we decided to just add a comment and go back to fix those
things when we have more time left.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

* Fix unit test

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

* Fix unit tests

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

* Cleanup old functions that were migrated

Some functions were left behind after being migrated to a action class.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

* Removing unused code

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

---------

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>