Add actions report to be displayed by the end of the analysis
This commit introduces the report summary that will be displayed to the
user by end of the analysis, regardless if they are using doing a normal
conversion, or a pre conversion check.

The report is pretty simple as it consist mainly of logging the message
to the user in a nice format, ordered by most important first.

The catch here is that if the user is doing a pre conversion analysis,
they will get all messages generated by the action classes displayed for
them, while, if they are doing a normal conversion, only `WARNINGS` and
above will be displayed.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>