Use format_report_message() to construct the logger.error messages.
The logger.error message is printed at the time an Action fails to give
the user immediate feedback (in case they want to hit Ctrl-C after
a failure).  Use the same function that report.summary() uses to
construct that message.

This also required moving the format_report_message and a few other
variables around to avoid circular imports.