Scale STATUS_CODE to the range 0-255
We may reuse STATUS_CODE for the program's exit code.  If we do so, we
will need the code to be in the range that's valid for exit codes on
Linux, (0-255).  Note that we've left room at the high end for any
status codes which are more severe than ERROR.