RHSM: do not use the 'force' D-Bus registration option
The 'force' D-Bus registration option used to simply be not implemented,
effectively being a no-op; because of that, the system is manually
unregistered before the actual registration.

Newer versions of subscription-manager (in 8.8 and 9.2) actually do
implement the 'force' option; OTOH, since the system was unregistered
before the registration attempt, having the option there does not make
any difference. Hence, simply drop the 'force' option altogether.

Improve the comment regarding the unregistration to mention the status
of the 'force' option, and the general reason for unconditionally
unregistering at that point.

Signed-off-by: Pino Toscano <ptoscano@redhat.com>