Move the determination of the activity into toolopts.
Right now we're using an environment variable to determine which activity convert2rhel is performing
but the plan is to use a subcommand on the CLI for the next release.  Move the code which decided
which activity we are performing into tool_opts now so that we only have to change the code there
when we make the switch.