Do not honor CONVERT2RHEL_UNSUPPORTED_INCOMPLETE_ROLLBACK when in analyze mode.
`analyze` depends on being able to rollback all changes to the system to get back to the initial
state.  If INCOMPLETE_ROLLBACK is specified, we wouldn't be able to do that, which would leave the
user with a system which can't be converted without manual intervention after running
`convert2rhel analyze.`  So we need to prevent that from happening.

Note that having the check for the environment variable and analyze activity in utils.download_pkg()
isn't great.  It means that we will report a FAILURE when we get there whereas it would be more
accurate to report a SKIP.  In the future we should redesign this so download_pkgs() raises an
exception in this case and a higher level catches the exception and decides what to do based on the
environment variable and the activity.