Expect libdnf-plugin-subscription-manager on RHEL9
The dnf-plugin-subscription-manager on RHEL8 got renamed to
libdnf-plugin-subscription-manager on RHEL9.

Also, the python3-syspurpose is no longer shipped on RHEL 9.

And, there are two subscription-manager-rhsm-certificates pkgs in UBI 9,
one for x86_64 and one noarch. yumdownloader downloads both for some reason so
we've now specified that we want the x86_64 one.