Fix for poor error message when a cert has already been removed in rollback.
If something (for instance, package removal) removes a certificate that we installed, then we will
try to reove the cert during rollback but the file won't exist. Add some code to give a nice message
in this case.

Also fix the integration test test_rhsm_error_logged().  It was checking for a prior bad error
message (an exception).  That won't happen in the same circumstance with the new code so adapt it to fit the new reality.