[RHELC-1107] Schema to validate the pre-conversion assessment json format  (#879)
* Add a schema for the 1.0 version of the assessment output.
* Add a variables field to ActionMessageBase.
Variables won't be implemented until the Q4 release but we can add an empty variables field now so
that the format remains the same.  This matches with the 1.0 assessment schema.
* Add variables parameter to the rest of actions.
add_message and some other methods were lacking the variables parameter,
this commit updates that and introduces that new parameter to the
action, defaulting to an empty dictionary.
* Hack together a separation of message into title, description, diagnosis, and remediation.
Before Q3 MVP, we want these fields to be passed in by the Action.  For now, to get the framework
updated and the assessment to validate with the schema, we will take the existing message and
attempt to set good values from it.
* Add test to verify the report json file existance
Verify the existence of the convert2rhel-pre-conversion.json and validate the schema of the report.
* The json output to emit empty string instead of None.
Found places where default values for the message and result fields was set to None and change
those to empty string.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>
Signed-off-by: Daniel Diblik <ddiblik@redhat.com>
Co-authored-by: Rodolfo Olivieri <rolivier@redhat.com>
Co-authored-by: Daniel Diblik <ddiblik@redhat.com>