Fix package backup on a pre-registered system
When the system is registered to RHSM before running convert2rhel, the
redhat.repo file is not generated because of the missing RHSM product
cert on non-RHEL system. However convert2rhel installs this cert and
when we do a package backup, the subscription-manager yum plugin
generates the redhat.repo file. Having the RHEL repos defined and
enabled during the package backup causes a failure when accessing the
RHEL repos due to the $releasever being different on RHEL and
RHEL-clones. Besides that, when backing up packages we want to
be accessing only the RHEL-clone repos, not RHEL repos.

Also, fix the check whether the packages to be removed have been really
removed. We were comparing strings to objects which led to a false flag
warning that packages that were in fact removed haven't been removed.

Also, update dependencies of a few Actions to make sure:
- we backup system repos before pkg backup because we're
  using the backed up repos for the pkg backups
- we remove pkgs that contain repofiles only after installing
  sub-man pkgs because the sub-man pkg installation may require
  dependencies from the original vendor repos