Fix sub-man installation on Alma and Rocky 8
If the python3-syspurpose package is installed on the system prior to
the conversion, it has a different NEVRA than subscription-manager
requires. The subscription-manager is strict about it - the syspurpose
NEVRA needs to match exactly the sub-man NEVRA. And because we're
downloading the sub-man packages from UBI, the NEVRA doesn't match.

Previously we thought this issue affects just CentOS Linux 8 but it does
affect Alma and Rocky 8 as well.