Update convert2rhel manpage
Generated a new manapge for convert2rhel that includes the analyze
subcommand, as well, the options that were changed in the last release.

In this commit, there is also some changes related to the toolopts.py
module that ensures the subcommands are displayed correctly in the
manpage and in with --help.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>