Update remediation key to match leapp report metadata
The remediation key was singular in our report, and leapp one is in
plural. We are matching the key in our report to remove this workaround
from the pre-conversion script that is executed with the
rhc-worker-script.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

Update remediation key in actions and tests

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

Update more remediation keys

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

Update remaining remediation keys

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>

Fix keys in leftover files after rebase

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>