Fix backtrace in kernel signature verification
The old method we used was causing a backtrace during the fingerprint
verification as it was consulting the yumdb and in return it was
reporting metadata that was not relevant.

This commit introduces an minimal refactor that changes the way we query
for the package, so instead of relying in querying the rpmdb first, we
use rpm directly to do the job.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>