Check for firewalld system unit
Firewalld is having problems with the python-nftables package after we
repalce the packages in the system to their rhel counterparts. This
commit introduces a way to check if firewalld is available and running,
and if so, we ask the user to stop it and re-enable after the
conversion.

If firewalld is left running during the conversion, the user will
probably see a set of errors happening in the systemctl logs for
firewalld, as well, they will be unable to ssh into the machine after
the conversion is done.

Signed-off-by: Rodolfo Olivieri <rolivier@redhat.com>